requirements.txt
langchain
langchain-community
streamlit
python-dotenv
langchain-chroma
langchain-groq
langchain-huggingface
pypdf
chromadb